Buttons Posts moved to NavigationBarController, made a InfoViewController with button See the post
