+++ Navigation/FeedViewController.swift
@@ -15,15 +15,9 @@ class FeedViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-    let button = UIButton(frame: CGRect(x: 100, y: 300, width: 200, height: 50))
-        button.setTitle("See the Post!", for: .normal)
-        button.addTarget(self, action: #selector(tap), for: .touchUpInside)
-        button.backgroundColor = .darkGray
-        button.setTitleColor(.black, for: .normal)
-        
-        view.addSubview(button)
-        
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Posts", style: .plain, target: self, action: #selector(tap))
+    
+    
     }
     
     @objc func tap() {
@@ -31,10 +25,6 @@ class FeedViewController: UIViewController {
         navigationController?.pushViewController(vc, animated: true)
     }
     
-    let alert = UIAlertController(title: "Error", message: "AZAZA", preferredStyle: .alert)
-    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel) {
-        _ in print("Cancel")
-    }
     
     var firstPost = Post(title: "FirstPost")
